+++ CHANGES.rst
@@ -80,7 +80,7 @@ Unreleased
     ``Retry-After`` header. :issue:`1657`
 -   ``Map`` and ``Rule`` have a ``merge_slashes`` option to collapse
     multiple slashes into one, similar to how many HTTP servers behave.
-    This is enabled by default. :pr:`1286`
+    This is enabled by default. :pr:`1286, 1694`
 -   Add HTTP 103, 208, 306, 425, 506, 508, and 511 to the list of status
     codes. :pr:`1678`
 -   Add an ``update`` method to the ``Headers`` data structure.