+++ CHANGES.rst
@@ -3,7 +3,7 @@
 Version 0.16.1
 --------------
 
-Unreleased
+Released 2020-01-27
 
 -   Fix import location in deprecation messages for subpackages.
     :issue:`1663`